// src/app/enums/skill-key.enum.ts
export enum SkillKey {
    ACROBATICS = 'ACROBATICS',
    ATHLETICS = 'ATHLETICS',
    BLUFF = 'BLUFF',
    COMPUTERS = 'COMPUTERS',
    CULTURE = 'CULTURE',
    DIPLOMACY = 'DIPLOMACY',
    DISGUISE = 'DISGUISE',
    ENGINEERING = 'ENGINEERING',
    INTIMIDATE = 'INTIMIDATE',
    LIFESCIENCE = 'LIFESCIENCE',
    MEDICINE = 'MEDICINE',
    MYSTICISM = 'MYSTICISM',
    PERCEPTION = 'PERCEPTION',
    PHYSICALSCIENCE = 'PHYSICALSCIENCE',
    PILOTING = 'PILOTING',
    // PROFESSIONS = 'PROFESSIONS',
    SENSEMOTIVE = 'SENSEMOTIVE',
    SLEIGHTOFHAND = 'SLEIGHTOFHAND',
    STEALTH = 'STEALTH',
    SURVIVAL = 'SURVIVAL'
}